Refactoring RootZoneState
